+++ FrontEnd/ContactPage/Contact.js
@@ -1,0 +1,15 @@
+
+import React from 'react';
+import HeadTitle from '../../common/HeadTitle/HeadTitle';
+import ContactForm from './ContactForm';
+
+const Contact = () => {
+  return (
+	<>
+	<HeadTitle/>
+	<ContactForm/>
+	</>
+  );
+}
+
+export default Contact;
